requirements.txt
bcc==0.18.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
diskcache==5.6.3
distlib==0.3.7
distro==1.7.0
filelock==3.13.1
Flask==3.0.0
greenlet==1.1.2
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama_cpp_python==0.2.18
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Markdown==3.5.1
MarkupSafe==2.1.3
more-itertools==8.10.0
msgpack==1.0.3
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
pandas==2.1.2
platformdirs==3.11.0
pop-transition==1.1.2
protobuf==3.12.4
psutil==5.9.0
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.31.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
soupsieve==2.5
systemd-python==234
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
virtualenv==20.24.6
wadllib==1.3.6
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
